requirements.txt
Django>=5.0
python-dotenv
pandas
openpyxl
python-docx
PyPDF2
tabulate
